commented main method and neatened up code

+++ Ingredient.cs
@@ -10,13 +10,13 @@ namespace ST10257503.PROG6221POEPart1
     {
 
         public string Name { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Unit { get; set; }
 
 
         public string Addition { get; set; }
 
-        public Ingredient(string name, int quantity, string unit , string addition )
+        public Ingredient(string name, double quantity, string unit , string addition )
         {
             Name = name;
             Quantity = quantity;
